requirements.txt
# ============================================
# KARI Dashboard - Python Dependencies
# ============================================
#
# Установка: pip install -r requirements.txt --break-system-packages
# (флаг --break-system-packages нужен для Windows без venv)
#

# ============================================
# DATA PROCESSING
# ============================================

# Работа с Excel файлами
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.0,<4.0.0

# ============================================
# TELEGRAM BOT
# ============================================

# Telegram Bot API (v20+ с async support)
python-telegram-bot>=20.0,<21.0

# ============================================
# CONFIGURATION
# ============================================

# Загрузка переменных окружения из .env
python-dotenv>=1.0.0,<2.0.0

# ============================================
# OPTIONAL (для разработки)
# ============================================

# Тестирование (раскомментируй если нужно)
# pytest>=7.0.0,<8.0.0
# pytest-asyncio>=0.21.0,<1.0.0

# Линтер (раскомментируй если нужно)
# flake8>=6.0.0,<7.0.0
